+++ domain_attribute.go
@@ -10,7 +10,7 @@ type DomainAttribute struct {
 }
 
 func (da DomainAttribute) path() string {
-	return "domain_attribute/" + strconv.FormatInt(int64(da.Id), 10)
+	return "domain_attribute/" + strconv.FormatInt(da.Id, 10)
 }
 
 type domainAttributeResponse struct {
